Fixed 'set mark'-bug and added etype-lines.el
Some functions that should only be used interactively has been
replaced. Added a etype-lines.el file with the default set of words, and
made a customizable variable that can specify the location of a file
containing lines of words.

Also I have started adding waves of words instead of a constant stream,
but this does not work well yet.